pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "translate-agent"
version = "0.1.0"
description = "Pipeline to translate English videos into Chinese dubbed videos with bilingual subtitles."
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Codex Assistant"}]
dependencies = [
    "openai>=1.14.0",
    "python-dotenv>=1.0.1",
    "moviepy>=1.0.3",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "numpy>=1.23",
    "torch>=2.0.0",
    "openai-whisper>=20231117",
    "tqdm>=4.66.1",
    "srt>=3.5.3",
    "requests>=2.31.0",
    "edge-tts>=6.1.10"
]

[project.scripts]
translate-video = "scripts.translate_video:main"
